LayoutTests/imported/w3c/web-platform-tests/fetch/metadata/fetch-preflight.https.sub.any.worker-expected.txt

CONSOLE MESSAGE: Preflight response is not successful. Status code: 403
CONSOLE MESSAGE: Preflight response is not successful. Status code: 403

FAIL Same-site fetch with preflight promise_test: Unhandled rejection with value: object "TypeError: Load failed"
FAIL Cross-site fetch with preflight promise_test: Unhandled rejection with value: object "TypeError: Load failed"
